drop down lists
Added some extra functions to pass information from the drop down
menus. It’s super sloppy and inefficient right now and needs to be
tightened up. Also, there’s a strange bug where once an option is
selected, it doesn’t update in the drop down but the value gets passed.
It could be a browser thing.